数据库访问接口

+++ hera-common/src/main/java/com/dfire/common/entity/HeraHostRelation.java
@@ -16,7 +16,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class HeraHostRelation {
 
-    private Long id;
+    private int id;
+
     private String host;
-    private Integer hostGroupId;
+
+    private int hostGroupId;
 }
